pyproject.toml
[tool.poetry]
name = "\\"
version = "0.1.0"
description = "4cid bot for LXD"
authors = ["rezvaya <rezvaya2016@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.1.1"
pytelegrambotapi = "^4.29.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
